+++ src/codemagic/tools/_app_store_connect/action_groups/builds_action_group.py
@@ -105,14 +105,18 @@ class BuildsActionGroup(AbstractBaseAction, metaclass=ABCMeta):
         else:
             max_processing_minutes = Types.MaxBuildProcessingWait.default_value
 
+        self.logger.info(Colors.BLUE('\nSubmit uploaded build to TestFlight beta review'))
+
         build, app = self.api_client.builds.read_with_include(build_id, App)
 
-        self._assert_app_has_testflight_information(app)
+        try:
+            self._assert_app_has_testflight_information(app)
+        except ValueError as ve:
+            raise AppStoreConnectError(str(ve)) from ve
 
         if max_processing_minutes:
             build = self._wait_until_build_is_processed(build, max_processing_minutes)
 
-        self.logger.info(Colors.BLUE('\nSubmit uploaded build to TestFlight beta review'))
         self.create_beta_app_review_submission(build.id)
 
     def _wait_until_build_is_processed(